[tests] add integration test for auth against cega
