+++ .github/integration/tests/sda/99_auth_test.sh
@@ -1,0 +1,9 @@
+#!/bin/sh
+set -e
+
+python -m pip install --upgrade pip
+pip install tox
+
+tox -e unit_tests -c /tests/sda/auth/tox.ini
+
+echo "auth test completes successfully"
